Added a m_prev_activations matrix and view to the layer class.  This is local "copy" of the error signal from "previous" layer and replaces the adhoc usage of the X and X_v matrices in the fully connected and softmax layer.  m_prev_activations is setup to point to the fp_input in the standard layer forward pass routine.  Note that this gets rid of the use of the Elemental Proxy readers and relies on the assignment operator= function to properly reshape data into the necessary target distribution.
Updated the categorical target layer classes to locally compute the
error function that is used by the "previous" layer during backprop.
Now it will subtract the ground truth labels from the the activations
of the forward prop "previous" layer, which corresponds to the
classifiers predicted classes.  This means that the softmax layer no
longer needs to perform this calculation.  Additionally, it means that
the softmax layer can no longer compute the cross entropy cost/loss
function.

Moved the computation of the cross entropy cost function out of the
softmax layer class into the target layer class.  This decouples the
cost/loss/objective function from the softmax layer and is the first
step to moving the objective function into its own class.