<lbann>/utility is a new directory that is intended to contain drivers (similar to <lbann>/model_zoo), however, these drivers perform support functions, e.g: lbann_compute_mean_data.cpp computes and writes to file mean data for a set of imagenet files.
Files in this directory, as in <lbann>/model_zoo, contain main().
This differs from src/util, which contains utility function only.

Note: although I'm pushing this commit, DO NOT attempt to compile
or run. This is because (1) the driver depends on modifications
to src/data_drivers/lbann_data_reader_imagenet.[hpp|cpp] which
are still in development; (2) support for compiling the driver
is not yet in our cmake infrastructure (because, um, I need help
getting that to work ...).