Fix Elemental build with CUDA
The libEl.so library was not given a library to resolve CUDA-RT symbols
when linked. This is not a problem when it is eventually used to link an
executable with a shared library version of the CUDA RT, but fails if
the static version is used. Since newer version of CMake uses the static CUDA RT library by default, the final linking of executables using libEl fails. This fix should force the libEl.so to look for the shared CUDA library even with the main executable is built with the static library.