Werror works as expected now
CMake's checks should not be required to be warning-free. However,
adding Werror directly to the CMAKE_CXX_FLAGS was causing some of the
required checks to fail, so it was claiming, for example, that MPI did
not work. It does work, but the check generates some warning (not
actually about MPI) which gets promoted to an error and the check
fails.

The flag is now added to LBANN as a (public) imported interface
library (imported because nothing is built; interface because
"-Werror" does not have a location on disk). In the future, this
should be a privately linked library so that it does not propagate to
unsuspecting downstreams who would then be subjected to the evils of
-Werror.