Multiple bug fixes for the distributed I/O buffer:
*Fixed a bug there the mini-batch size would be different for multiple
models on the last round when model 0 had an extra sample.

*Fixed a bug where the distributed I/O buffer always assumed that rank
0 of model 0 had the extra sample on a partial round.

*Fixed a bug where the combination of the base and model offset were
incorrect for multiple model when there was so little data that a
ranks only mini-batch was a partial mini-batch (and thus the lower
rank models did not complete a full mini-batch for that round either).

*Fixed a bug where a null pointer was dereferenced in the callback for
printing if there was no valid testing data.

*Changed the reporting for the number of mini-batches and size of the
last mini-batch to include the world remainder data if is owned by a
rank other than the world master.