Always check CUDNN return values
Checking return values is not a performance concern, so it should always
be done. Previously, CHECK_CUDNN only checks errors with Debug build,
but that also introduces fairly heavy-weight device synchronizations
before and after error checking. FORCE_CHECK_CUDNN always checks return
values irrespective of the build mode, but due to the synchronizations
it is not only too costly but also can cause deadlocks.

This PR changes the behavior such that CUDNN return values are always
checked in any build mode. It still keeps the synchronizations for the
Debug build mode, but in my opinion we should also remove them as they
make the macro unusable.

FORCE_CHECK_CUDNN is not necessary anymore as CHECK_CUDNN always checks
errors. It is still there in order to make this PR minimal.

If this looks good, CUDA_CHECK and FORCE_CHECK_CUDA should be similarly
changed too.