Introduce rank stride for better mapping file loading ranks over nodes
When the mini-batch size is smaller than the number of ranks, LBANN uses
only the first ranks of the mini-batch size to read data when the input
layer is data-parallel format. This is inefficient when followed by a
convolution with spatial partitioning as it needs to shuffle data across
nodes. With rank stride, MPI ranks are reordered with that stride. So,
if a node has 4 ranks and the rank stride is configured to be 4, only
the rank 0 at each node is used to load data. Since this reorders rank
and node mapping, it may negatively impact other LBANN layers. Distconv
still uses the original rank ordering, so it should not be affected.