Cleaned up and properly segmented the checkpoint and restart logic across the trainer, execution context, and model.  Additionally, pushed the execution context through more of the code.
Moved setup of the data readers from the input layer into the
lbann_library build_model_from_prototext function.  This alleviates
the need for the model to have direct access to the trainer.  Fixed
several bugs during setup where the model does not have a valid
execution context.  Moved the reporting of most of the LBANN
configuration and run state so that it is reported by the trainer, not
the model.

Fixed a bug where the data store was accessing the execution context
during setup.

Cleaned up some C++ style and improved use of unique pointers.