Hacked bugfix for sendrecv bug in MVAPICH2-2.3.
MPI_Sendrecv will output all zeros if two processes are on different nodes, the data buffers are in pinned memory, and the buffer size is between approximately 16 KiB and 16 MiB. This caused problems in LTFB since AlexNet convolution kernels fit within this size range. This bug does't show up in MVAPICH2-2.2 or MVAPICH2-2.3.1. A bug report has been submitted to LC and they are forwarding it to the MVAPICH folks.