Incomplete first sweep at callback re-factory-ing.
This will compile and run for all but a few callbacks. The old factory
function used a "layer_list" and a "weights_list" argument to select
objects that would then be passed into the model. This detail has been
glossed over for now as it is logically inconsistent (these objects
should be grabbed in the context of a model, i.e., during setup() or
just-in-time in the callback functions themselves. The fix for this
will be coming soon.