Factoring out the trainer and training state from the model (#916)
Refactored LBANN into trainers, execution contexts, training algorithms, and model classes.  There are additional improvements that crept into the PR that include fixes to the checkpoint and restart, as well as introducing the use of the Cereal library into the code to replace behavior that was in the persist class.

The details of these new classes is listed below:

Trainers (i.e. execution environment)
******************************************

A trainer is a collection of compute resources and defines a explicit
communication domain.  It provides the execution for both the training
and inference of a trained model.  Once constructed a trainer owns an
LBANN comm object that defines both intra- and inter-trainer
communication domains.  Additionally, a trainer will contain an I/O
thread pool that is used to fetch and pre-process data that will be
provided to the trainer's models.

A trainer owns:

* comm object
* I/O thread pool
* One or more models
* Execution context for each model
* In the future, it will also contain the data readers.

Execution Context
******************************************

When a model is attached to a trainer the execution context of the
training algorithm is stored in an execution_context class (or
sub-class) per execution mode.  Thus there is one execution context
per model and mode that contains all of the state with respect to the
training algorithm being applied to the model.

For example it tracks the current:

* step
* execution mode
* epoch
* and a pointer back to the trainer

Termination Criteria (Pending)
******************************************

(Pending feature) When a model is going to be trained or evaluated,
the termination criteria is specified in an object that is passed into
the training algorithm.  (Note that this feature is under development,
currently the termination criteria is dictated by when the training
algorithm executes a fixed number of epochs.)

Training Algorithms
******************************************

The training algorithm defines the optimization that is to be
applied to the model(s) being trained.  Additionally, it can
specify how to evaluate the model.

Model
******************************************

A model is a collection of operations with dependencies encoded as a
directed acyclic graph (DAG).  In a typical formulation, these
operations form a neural network that will be either trained or used
for inference.  Each operation in the model is an instance of the
layer class.  The model is then a collection of layers that perform
transformations and mathematical operations on data that is passed
between layers.  The model's DAG is executed in topological order.
Inside of some layer types are weight matrices that define a trained
model.  (Note that LBANN should be able to support non-DNN models, but
this is a subject for future work.)

Each layer in the graph contains a set of tensors that holds the
inputs, computed outputs, gradients with respect to the outputs, and
gradients with respect to the inputs.  Furthermore, for each layer in
the graph with learnable parameters, there is an associated weight
tensor that form the learned weights of the model.  The model also
owns the objective function, since that is integrally tied into the
model's computational graph.  Additionally, the model owns both the
default optimizer that is used to provide a standard optimizer for the
model's weight tensors.  Once each weight tensor is instantiated, it
will owns an instance of an optimizer.

The model also owns the max_mini_batch_size that is supported by the
model.  This is due to the fact that it changes the size and shape of
input, output, and gradient tensors.  Additionally, the model owns a
field that controls if background I/O is allowed for this model and
associated data reader.

******************************************

* Cleaned up and properly segmented the checkpoint and restart logic
across the trainer, execution context, and model.  Additionally,
pushed the execution context through more of the code.

Moved setup of the data readers from the input layer into the
lbann_library build_model_from_prototext function.  This alleviates
the need for the model to have direct access to the trainer.  Fixed
several bugs during setup where the model does not have a valid
execution context.  Moved the reporting of most of the LBANN
configuration and run state so that it is reported by the trainer, not
the model.

Cleaned up some C++ style and improved use of unique pointers.

* Updated the model prototext files to include the trainer message.
Added a trainer prototext message to the python interface.

* Changed the model factory to return a unique_ptr.  Clarified the
documentation of the step field in the execution_context.

* Cleaning up what gets done in the model's checkpoint and restart
versus the execution context.  Note that the model has to be able to
be restored without an execution context since it can be loaded from a
trained model or a checkpoint.

* - Updated some of the classes to use the Cereal library to save their
  checkpoint data.  For these classes, data is now stored in XML
  format. Supported classes are:
  - Data reader
  - Execution context and subclasses
  - Metrics
  - Generic input layer
- Removed explicit packing and unpacking routines that have been
  replaced with Cereal
- Added templated functions for writing out Cereal archives to XML
  file or to a binary buffer that can be exchanged between ranks
- Added support for how the new execution modes interact with the
  execution contexts
- Added a callback phase that the checkpoint callback uses to identify
  the state at which a helper function was called
- Improved the naming convention for how checkpoints are
  differentiated based on trainer, model, execution mode, and epoch
  plus step
- Changed the checkpointing callback so that all of the state is saved
  at each checkpoint time.  Added support for explicitly saving the
  model weights, the trainer state only (execution contexts), or a
  full checkpoint
- Added support to the trainer to create and cache new execution
  contexts

* Updated mini-batch size, step, and epoch counters to use size_t instead of int

* Changed the error reporting mechanism for the model to save and
restart, such that if a layer fails it will throw an exception.  Also
renamed the RNG save and load functions to indicate that it is for all ranks.

* Simplified how the checkpoint directory is communicated for shared
checkpoints to address racy bug in saving the RNG.

* Added error handling for the RNG checkpoint and restart

* Added a future clarification for how to restart from a checkpoint.

* Added a flag to allow the checkpoint directory to be set from the
command line.  Updated the unit checkpoint tests to use explicit
directories to avoid moving checkpoints after the fact.

* Fixed a bug in proto_common that failed to detect when a directory
doesn't exist.  Added support to override the directory used by the
dump weights callbsck and save model.  Fixed a bug in the front end
codes where the protobuf for the models was using a copy instead of
the pointer.