Refactor weights and objective fns to get comm at setup (#1247)
* Created a private default constructor for the weight class that does not take 
the comm object.  This allows the deserialization to create new instances
of the weight class that can then complete their initialization in the restart
logic.  This will make it easier to restore weight classes from archive files.

*  Moved the initialization of the weights comm object and default matrix
distribution into a separate function and out of the constructor.

* Updated the optimizer classes to get the comm object during setup
rather than construction.  This is necessary for using Cereal for
checkpoint and restart of the classes.

* Changed get_comm in the weights class to return a reference.