Fix hang in HDF5 MPI-IO
HDF5 caused hanging. Likely because a HDF5 property was created with MPI
at every fetch_datum. The property is now moved out of the function and
is only done once, so it should not hang anymore. Yet, MPI-IO is disabled for now. Should be looked into again once everything becomes working.