Adding the data-store capability for the Cosmoflow HDF data reader (#93)
* Added support for placing a slab of a sample's data into the data
store.  Then each partition of data (slab) is manged as if it were a
subsample and exchanges within a fixed offset of nodes.  Also added
support in the data store to record the sample ID and slab ID.

* Testing alternate configurations of world rank.

* Fixed a performance bug where a conduit node was being copied instead
of getting an external reference.

* Fixed a second performance bug where a new conduit node was being
passed in to the function reading the hyperslab of data, and then it
was copied outside.  Also reverted the calculation of world rank to
use the distconv methods.

* Fixed a bug where the H5Dread command was not passed the data transfer
properties.  Switched to using H5FD_MPIO_INDEPENDENT instead of
H5FD_MPIO_COLLECTIVE as the default.

* Add nvprof makers

* Minor cleanup

* formatting

* Fix protobuf version in superbuild

* Cleaning up

* Make use of data store optional in the HDF5 data reader for Cosmoflow

Use --use_data_store when running lbann to enable data store

* Cleaning up HDF5 data reader