Refactor add data type layer (#1261)
* Adds a TensorDataType template parameter to the layer, weights, and
optimizer classes as appropriate.  This minimizes the dependency of the
code on the global macro definition of type DataType.  This is the first step
in a sequence to enable LBANN to have per-layer data type support.

* Refactoring the layer class hierarchy to be templated by a
TensorDataType.  Added a data_type_layer in-between the Layer class
and all children.  The data_type_layer holds the activations, error,
and weight matrices.  It is the first class in the hierarchy that is
specialized by TensorDataType.

* Moved all of the TensorDataType dependencies in the weights class into
a data_type_weights subclass.  This keeps weighs a non-templated class
again and should make it possible for the model to maintain a list of
weights without know which data type each one is.

* Create a data_type_weights_initializer subclass

* Moved all of the TensorDataType dependencies in the optimizers class into
a data_type_optimizers subclass.

* Minor cleanup and clang -g fix for all files on os x.

* Builds with DataType equal to float or double.