Implement ETI for multiple data types (#1361)
* Added macros for explicitly instantiating classes and functions over multiple 
  data and device types

* Extended data_type_layer and fully_connected_layer to work with both float
  and double types.

* Updated data_type_weights to support double precision.  Added
templating to the summarizer so that every layer type can invoke the
summarizer at the proper precision.  Internally the summarizer stores
scalars and floats and histograms as doubles based on the TensorBoard
interface.  Added templated functions to both the comm class allreduce
functions and the persist classes read and write matrix functions.

* Updated weight initializers, persist class, and random initializers to
allow templating for the tensor data type.

* Updated comm class for ETI with templated matrix allreduce operations.

* Updated ETI for data_type_optimizer.  Added templating for more
functions in persist class and fixed inline bug in summarizer.

* Fixed a bug in the computation of the histogram for the summarizer
where the bucket index goes out of bounds.

* Added versions of all Hydrogen matrix layouts templated over TensorDataType.