Updating the layer factories to provide ETI for float and double. (#1370)
* Updating the layer construction to provide support for layers that have 
ETI for float and double and starting to migrate to a factory based 
implementation.

* Added support for changing the type of fully connected and convolutional
layers.

* Added macro to disable compiling with double precision with the
instantiation macro.

* Adding a new factory method for instantiating layers that support
multiple data types and allow the prototext to select the data type.

* Updated the python front end to take the datatype argument.

* Added a utility function to get the string of a data type

* Exposed the reconcile_values function in the base weights class.
Fixed a bug in the type specific implementation of the
reconcile_values function.  Updated base Hydrogen version to 1.3.2 to
access bugfixes for on device data conversion.