Feature supporting half datatype (#1378)
* Adding support for a CPU Half type (cpu_fp16) and a GPU Half type
(fp16).  Added the C++ header only library to the build environment
for the spack build.

* Switching to using Hydrogen trait types.

* Fixed the scoping of ETI in persist

* Corrected the way that instantiate_device.hpp works to properly
instantiate the correct half data type for CPU and GPU devices.
Updating the setting of constant values to work with the Half cpu
header library.

* Explicit type conversion is required when dividing a half by an
integer.

* Update the random gaussian fill to use a float for generating the numbers when using cpu_fp16 data type.

* Changing so that we convert operands to the correct type as early as
possible so that as much arithmetic as possible is done in the layer's
native data type.

* Removed the explicit instantiation of the GPU and CPU half data types
when using the instantiate_device.hpp macro.  These definitions are
provided in the macro header.

* Switched over to using Hydrogen versions of sqrt and pow

* add greater support for __half in cuda basic functions

* instantiate data_type_layer with gpu_half

* use instantiate infrastructure for unary/binary layers

* Switching to using hydrogen and half versions of std functions

* add gpu half type instantiation to data type optimizer

* non-lbann_comm link-time fixes

* clean up comm for better integration with half types

* Fixing bugs in the base convolutional layer and cudnn utils functions
with respect to DataType.

* Added support for explicitly instantiating weights and initializers in
specific data types.

* cleanup things in the weights factory

* instantiate optimizers of the right types

* add temporary h2 implementation of switch-on-type dispatch

* clean up typelist stuff to use h2 typelist instead of el typelist

* Updated the check_gradient callback to work with weights of difference
data type.

* Added instantiation for FP16 types

* For the integrity of the check gradient test, the current
implementation uses an epsilon based on the minimum step size of the
float data type.

* Fix incorrect type for indices in evaluation layer

* Added a special data type for the scaling factors that are fed into
cuDNN calls.  Turns out that for fp16 data types, the scaling factors
still have to be float.  If they are passed in as half, no errors are
thrown.

* Added appropriate guards to protect FP16 and cpu_FP16 data types.

* Updated the version of protobuf being used and fixed the spack
environment so that it forces py-protobuf and protobuf to the same
version number.  Added a spack environment for non-GPU enabled systems.