Feature cleanup spack env (#1414)
* Simplified and updated the way that LBANN uses Spack to install
dependencies.  Revised installation instructions for both users and
developers.  Switched over to using named Spack environments for
installation of both the user install and installation of dependencies.
User primary interacts with an install_lbann.sh script and a
build_lbann_from_source.sh script.

* Moving packages that are used for auxiliary scripts out of the
mainline spack build environment into a separate build environment.

* Updated the OSX build to Mojave.

* More cleanup and rearrangement of the spack based build environment.
Added the preferred compiler to the list of spack dependencies so that
it is properly included in the module list.

* Adding external packages for NERSC

* Added support for making sure that the modules are properly setup for
the compute nodes at NERSC.

* Removed the yaml files in favor of the HERE based bash shell files for
defining the spack packages. 

* Updated the documentation for building LBANN as a user and on OSX.

* Separated the building of dependencies with spack and the
actual build of LBANN for developers.

* Updated the install script to explicitly define the dependencies since
spack is unable to skip modules that it cannot load.

* Added an option for superbuilding LBANN plus Hydrogen and Aluminum.

* Update the auxiliary tools to the same script format.

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>