Feature move checkpoint to trainer (#1457)
* Moved the checkpoint callback to be defined in the trainer's scope.
Added a callback list to the trainer class and updated both the model
and trainer to use list of shared pointers.

* Updated the python front end to support the trainer receiving callbacks

* Moving the checkpoint callback to the trainer class

* Moved the save and restore of the execution contexts into the trainer
class.  Updated the checkpoint and restart logic to place the trainer
namespace before the model and to capture all models inside of the
named trainer's directory.  As a result of this the callbacks setup
function for the model also requires a trainer's name to properly
setup the model.  Added support for the model save and restore
functions to specify a path that includes the trainer's name.

* During the setup call the trainer will now set the active trainer name
so that when a model is being setup, it can pull the trainer name from
the class.