Bugfix checkpoint trainer (#1480)
* Fixed a bug where the metrics were not properly saving the fields
from the child layer_metrics class.  This required pushing the actual
save and load functions to the child class.

* Fixed bug in the build script to reflect that the superbuild is not
used by the build script any more.

* Updated the checkpoint and restart logic so that a checkpoint is
properly encapsulated under a training algorithm at a specific point
in time.  This means that the checkpoint hierarchy is specified as the
trainer, then the training algorithm at a particular epoch and step,
and finally all of the appropriate models.  Moved the state of the
random number generators to the trainer / training algorithm level and
out of the model.

* Added some checking to make sure that the checkpoint directory exists

* Fixed dump weights to use the model's name

* Updated the Python tools testing framework to delete the work
directory before running the experiment.  Added the ability to setup a
work subdirectory.

* Fixed the restart_dir flag so that it applies to the checkpoint and
restart callback as indicated in the description.  Now it will allow
the checkpoint callback to restart from a different directory than the
one to which it will checkpoint.  Added return codes to the functions
that search for and open checkpoints.  Added some additional warning
messages if a valid checkpoint cannot be opened.

* Corrected support for the restart_dir flag to be directed to the
checkpoint callback.  Added support for the load_model_dir flag to
provide command line override for the load_model callback.

* Created a new test for the checkpoint and restart functionality that
is using the Python front-end.

Updated the trainer load from checkpoint functions to propely return a
boolean flag.  Added a LOG_MSG function to give better logging output.

* Pushed some functions used to check the results of the checkpoint and
restart test into the bamboo tools scope so they can be reused across
multiple tests.

* Updated the execution contexts to use references to the trainer and
training algorithm.

* Updated the training algorithm function to take a vector of models.
Renamed the flag in the model class to be more precise that it records
if a model has been setup.

* Changed the save_model callback to write to a directory structure that
is only specified by the trainer and model names, not the epoch and
step.

* Added another callback phase to the persist class to allow the
save_model callback to only save the model's weights and not
associated optimizer state.  Additionally, added a method for creating
the persist directoy independently from setting up the metadata files.

* Replaced the lbann2 reload unit test with a new one based on the
python front end that compares the result of loading weights from a
checkpoint to loading weights from a save_model callback.  Deleted the
lbann2 front end since its key purpose was to transfer weights between
two models which is now accomplished by either the save/load model
callback or the replace_weights callback.

Co-authored-by: Tom Benson <benson31@llnl.gov>
Co-authored-by: Tim Moon <moon13@llnl.gov>