Cleanup checkpoint and restart hierarchy (#1490)
* Working on fixing the hierarchy of the checkpoints to make sure that
state is properly encapsulated in the right level of the model or
training algorithm.  Also pushing checkpoint and restart out of the
persist objects into Cereal.

* Migrating optimizers to use Cereal archives for the checkpoint and
restart logic.  Added fix from benson31 to provide shim for handling
CPU and GPU half data types.

* Cleaned up the calls to generate checkpoint names.

* Switched hypergradient adam and rmsprop to use Cereal.

* Optimizer checkpoint and restart moved entirely into Cereal archives

* Fixed up the persist class functions that write out Cereal archives to
allow direct filenames to be provided and to use checkpoint dir if no
persist type is provided.  Corrected behavior of the checkpoint
callback to properly write the model archive int the correct part of
the checkpoint hierarchy.

* Removing vestigial use of te persist library for saving state from
the model class.  Updated the metrics class to place the metrics
archive inside of the model.

* Added support for using Cereal to checkpoint and restart the persist
objects callback_type field.  Cleaned up the persist object to remove
the file descriptor map, and all functions associated with writing to
the legacy binary blobs that were previously used for checkpoint and
restart.  The persist object will no longer open any files directly.
The open_checkpoint call will now only set the explicit filenames if
they are not already initialized.

* Removed legacy storage of checkpoint and restart logic to persist
binary blobs in favor of Cereal serialization.

* Fixed a bug in the model and trainer construction so that the command
line arguments were applied during the trainer's construction rather
than the model's so that the trainer's initialization would see the
changes.

* Moved the persist object out of the checkpoint class and into the
trainer class to make the checkpoint class more stateless.  Updated
the trainer class to save and restore its state via Cereal
serialization and to properly use the setup phase for reloading the
trainer.  Updated the find and open latest checkpoint functions in the
checkpoint class to operate with strings rather then fetching the
trainer, execution context, training algorithm or model objects.
Updated the checkpoint and restart logic to properly find the latest
checkpoint for the trainer during setup.  Cheanged the optimizer class
to not save the step time field as it does not need to be restored.

* Added the model directory to the path when loading models

* Fixed integration with new serialized support

* Apply suggestions from code review

Code cleanup from benson31.

Co-Authored-By: Tom Benson <30674819+benson31@users.noreply.github.com>

* Apply suggestions from code review

More edits from benson31.

Co-Authored-By: Tom Benson <30674819+benson31@users.noreply.github.com>

* Cleaned up the interface for find_latest_checkpoint

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>