Bugfix trainer specific callback outputs (#1523)
* Added helper functions to return the state of an execution context as
a string.  This allows it to be easily used as part of a directory
hierarchy or file path for logging, debugging, etc.

* Created a trainer aware function that has the trainer master rank
create a directory and all other ranks just synchronize on it.

* Created helper functions in the callback class to allow callbacks to
create trainer and model aware directory hierarchies and to make it
easier to get execution context specific file names.  Setup the dump
outputs callback to use the multi-trainer aware directory structure.