Refactor data coordinator (#1446)
* Create a data coordinator that exists at the trainer level and is 
responsible for the data readers and interfacing with the training
algorithms execution contexts.  In this PR the functionality of the layers
is largely unchanged, and reaches back into the data coordinator to
access the data readers.

* Moving the data reader into the data coordinator.  This pulls it out
of the model and generic_input_layer.

* Moved the data reader out of the layer factories where they are
constructed.  Now when layers require information from the data reader
it is done at setup time.

* Fixing up data coordinator

* Working on the python front end

* Changed the setup methond in the layer class to take the max supported
mini-batch size as an argument to move towards a moe functional / pure
compute graph model representation.

* Updated the layer setup functions to take both the maximum mini-batch
size required for the model as well as the dimensions of the data
set.

Moved functionality from the partitioned_io_buffer into the
data_coordinator for calculating the number of steps in a data set and
the number of allowed parallel readers.

* The data coordinator now has an observing pointer to the active
execution context.  This allows the data readers to get access to the
currently executing state without going through the model.  Data
readers no longer have an observing pointer to the model class.
Updated the training algorithm to pass a data coordinator along with
the model throught the execution functions.

* Updated application python code to pass the max mini-batch size to the trainer not the model.

* Fixing up tests

* Fixing the python interface

* Fixed where the data coordinator is initialized.

* Fixed compiler errors

* Removed RNG checkpoint restore from the model

* Moved the setup of the data readers into the setup of the data
coordinator.  Reworked the lbann library so that the RNG is initalized
before the trainer is setup and possibly restored from checkpoint.
Moved the restore of the RNG and data coordinator to the trainers load
from checkpoint function that occurs during the model / trainer
restart, rather than the trainers restore method.  Still investigating
why it has to occur at this point.  With this fix, the PR is passing
the checkpoint test.

* Moved the setup of the data store into the data coordinator.  Also
setup the data coordinator prior to calling setup on the trainer's
callbacks, and thus prior to restoring from checkpoint.  Restoring of
the RNGs has to occur on restart, rather than during trainer reload.

* Deprecated and removed the ability to split a data set over multiple
trainers (formerly called models).  Going forward, in a multi-trainer
environment, each trainer needs an independent sample list.  Removed
associated field from the proto message of the input layer to select
between the modes.

* Removed first_model flag from the build model function

* Apply suggestions from code review

Added suggestions from Tim about PFE.

Co-Authored-By: Tim Moon <moon13@llnl.gov>

* Addressed Tim's comments on how to setup mini-batch size for the transformer.

* Added check for execution modes with no valid data readers.

* Fixed new tests in the PFE

* Moved the slice points functionality into the JAG Conduit data reader
and provided a generic implementation in the top level data reader.

* Removed copy/view matrix code that should have been removed during rebase.

* Updated the prototext to reflect that the trainer holds fields about
the data readers.

* Updated the model setup and layer setup functions to take a
data structure that includes all metadata from the data_coordinator
and data readers.  This now includes any potential slice points in
addition to the data dimensions.  This allows the slice layer to get
everything it needs for setup from its instantiation and setup
function arguments.

* Updated JAG Conduit data reader to ignore queries for a N/A slice point

* Update include/lbann/data_coordinator/data_coordinator.hpp

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Update include/lbann/data_coordinator/data_coordinator.hpp

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Address feedback from Tom

* Apply suggestions from code review

Co-authored-by: Tim Moon <moon13@llnl.gov>

* Added more edits from PR review

Co-authored-by: Tim Moon <moon13@llnl.gov>
Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>