Fix LBANN inference (#1549)
* Updated the load_model callback so that it works on inference only
runs.  Updated the lbann_inf front end so that it uses the
lbann_library support for the load_model callback, rather than an
explicit call to a static function.  Changed how the evaluate function
sets up the model and data coordinator so that it has the execution
context and mode at the start of the function. This fixes a bug where
a model is called in inference only.

* Updated the command line arguments for loading the pre-trained model.