Bugfix objective function initialization (#1554)
* Updated the pointer to the objective function in the model class to
use a unique pointer.

* Added back the call to setup the objective_function class to the model
setup function.  This bug was introduced in
52b26d0fbb45b7d70fff5627e40ebffd655847c7.  Also added CEREAL archive
functions to the objective_function class and objective_function_term
class hierarchy.

* Updated the evaluation layer to use EvalType for the internal local
values matrix.  This minimizes the inconsistency when the layer
accumulates the scalar value in the OpenMP parallel loop, which causes
unnecessary variation in the objective function value.  Changed the
objective function class to reset the statistics rather than deleting
entries from the map.

* Changed MNIST data reader to log to cout rather than cerr.

* Fixed type for fp_cpu and fp_gpu calls with half types

* Removed last instances of LBANN_SEQUENTIAL_INITIALIZATION from the
CMake infrastructure.  LBANN_SEQUENTIAL_INITIALIZATION was removed
from the code base in a prior commit.

* Updated the evaluation layer to compute and store its internal values
at either TensorDataType or EvalType based on the LBANN_DETERMINISTIC
compile time flag.  Added some Hydrogen helper functions to provide a
view or copy of a local Hydrogen matrix as appropriate, which helps
with the evaluation layers conversion of the input activations.
Updated the print_parameters function to report if deterministic
execution was enabled during compilation.

* Updated the bamboo tests to build with deterministic exection by
enabling LBANN_DETERMINISTIC at compile time.

* Added a function to manually diff two directories and allow for
individual secondary differences of files with discrepancies.  Added a
second function to diff two XML files and allow for small numeric
errors in floating point numbers.

* Updated the checkpoint functions to use new approximate diff utils.
Changed the tolerance to 1e-8 in the approximate diff function.

* Relaxed the tolerance for approximate equal on floats in XML files.

* Update CMakeLists.txt

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>

* Updated the approximate diff function to take the tolerance as an
argument and to report it in the near diff case.

* Updated the different lines to also report the tolerance

Co-authored-by: Tom Benson <30674819+benson31@users.noreply.github.com>